+++ packages/cli/src/lib/commands/startLocalDevelopmentServer/server.js
@@ -22,4 +22,5 @@ app.use(
     }
   })
 )
+
 module.exports = app