+++ packages/core/src/decorators.test.ts
@@ -34,7 +34,7 @@ describe('Intent decorator', () => {
 
     beforeEach(() => {
       fetch.resetMocks()
-      fetch.mockResponseOnce(JSON.stringify(collection))
+      fetch.mockResponseOnce(JSON.stringify({ data: collection }))
     })
 
     it('adds a method', () => {
@@ -42,12 +42,10 @@ describe('Intent decorator', () => {
     })
 
     it('calling methods return a promise', () => {
-      expect(decoratedInstance.getCollectionIntentProp().constructor).toBe(
-        Promise
-      )
+      expect(decoratedInstance.getCollectionIntentProp().constructor).toBe(Promise)
     })
 
-    it('uses GetResourceIntent', async () => {
+    it('uses GetCollectionIntent', async () => {
       const success = jest.fn()
 
       await decoratedInstance
@@ -56,7 +54,7 @@ describe('Intent decorator', () => {
         .catch(a => console.log(a))
 
       expect(fetch).toBeCalledWith(
-        'http://localhost:5555/api/v1/1234/getCollectionIntent?page=1&setupId=&configId=&integrationId=42',
+        'http://localhost:5555/api/v1/1234/getCollectionIntent?page=1&setupId=&integrationId=42',
         commonParams
       )
 
@@ -69,7 +67,7 @@ describe('Intent decorator', () => {
 
     beforeEach(() => {
       fetch.resetMocks()
-      fetch.mockResponseOnce(JSON.stringify(item))
+      fetch.mockResponseOnce(JSON.stringify({ data: item }))
     })
 
     it('adds a method', () => {
@@ -77,9 +75,7 @@ describe('Intent decorator', () => {
     })
 
     it('calling methods return a promise', () => {
-      expect(decoratedInstance.getResourceIntentProp().constructor).toBe(
-        Promise
-      )
+      expect(decoratedInstance.getResourceIntentProp().constructor).toBe(Promise)
     })
 
     it('uses GetResourceIntent', async () => {
@@ -91,12 +87,12 @@ describe('Intent decorator', () => {
         .catch(a => console.log(a))
 
       expect(fetch).toBeCalledWith(
-        'http://localhost:5555/api/v1/1234/getResourceIntent?setupId=&configId=&integrationId=42',
+        'http://localhost:5555/api/v1/1234/getResourceIntent?setupId=&integrationId=42',
 
         commonParams
       )
 
-      expect(success).toBeCalledWith({ object: item })
+      expect(success).toBeCalledWith({ object: item, referenceId: undefined })
     })
   })
 })