+++ packages/core/src/decorators.ts
@@ -26,9 +26,10 @@ export interface IBearerStateDecorator<T> {
   (options?: IBearerStateDecoratorOptions): T
 }
 
-export const BearerState: IBearerStateDecorator<any> = (
-  options?: IBearerStateDecoratorOptions
-) => (target: any, key: string): void => {}
+export const BearerState: IBearerStateDecorator<any> = (options?: IBearerStateDecoratorOptions) => (
+  target: any,
+  key: string
+): void => {}
 /**
  *  Component Decorator
  */