+++ packages/core/src/requests.ts
@@ -1,7 +1,7 @@
 /* global fetch */
 import Bearer from './Bearer'
 
-export function bearerRequest(uri: string, baseParams = {}) {
+export function bearerRequest(uri: string, baseParams = {}): (params: any, init?: any) => Promise<any> {
   const url = `${Bearer.config.integrationHost}api/v1/${uri}`
 
   return function(params = {}, init = {}) {
@@ -20,9 +20,7 @@ export function bearerRequest(uri: string, baseParams = {}) {
           integrationId: Bearer.config.integrationId
         }
 
-        const query = Object.keys(sentParams).map(key =>
-          [key, sentParams[key]].join('=')
-        )
+        const query = Object.keys(sentParams).map(key => [key, sentParams[key]].join('='))
         const uri = `${url}?${query.join('&')}`
         console.debug('[BEARER]', 'fetch', url, query)
         fetch
@@ -47,8 +45,16 @@ export function itemRequest() {
   return bearerRequest('items')
 }
 
-export function intentRequest({ intentName, scenarioId, setupId, configId }) {
-  return bearerRequest(`${scenarioId}/${intentName}`, { setupId, configId })
+export function intentRequest({
+  intentName,
+  scenarioId,
+  setupId
+}: {
+  intentName: string
+  scenarioId: string
+  setupId: string
+}) {
+  return bearerRequest(`${scenarioId}/${intentName}`, { setupId })
 }
 
 export default {