+++ packages/intents/src/index.ts
@@ -1,4 +1,4 @@
-import axios, { AxiosInstance } from 'axios'
+import axios, { AxiosInstance, AxiosResponse } from 'axios'
 
 import { sendSuccessMessage, sendErrorMessage } from './lambda'
 
@@ -7,13 +7,14 @@ export type TContext = {
   [key: string]: any
 }
 
+export type TStateData = AxiosResponse<{
+  Item: any
+}>
+
 export class Intent {
-  static getCollection(
-    callback,
-    { collection, error }: { collection?: any; error?: any }
-  ) {
+  static getCollection(callback, { collection, error }: { collection?: any; error?: any }) {
     if (collection) {
-      sendSuccessMessage(callback, collection)
+      sendSuccessMessage(callback, { data: collection })
     } else {
       sendErrorMessage(callback, { error: error || 'Unkown error' })
     }
@@ -21,7 +22,7 @@ export class Intent {
 
   static getObject(callback, { object, error }: { object?: any; error?: any }) {
     if (object) {
-      sendSuccessMessage(callback, object)
+      sendSuccessMessage(callback, { data: object })
     } else {
       sendErrorMessage(callback, { error: error || 'Unkown error' })
     }
@@ -39,17 +40,11 @@ export const STATE_CLIENT: AxiosInstance = axios.create({
 
 class BaseIntent {
   static get display(): string {
-    throw new Error(
-      'Extending class needs to implement `static intent(action)` method'
-    )
+    throw new Error('Extending class needs to implement `static intent(action)` method')
   }
 
-  static intent(
-    action
-  ): (event: any, context: any, callback: (...args: any[]) => any) => any {
-    throw new Error(
-      'Extending class needs to implement `static intent(action)` method'
-    )
+  static intent(action): (event: any, context: any, callback: (...args: any[]) => any) => any {
+    throw new Error('Extending class needs to implement `static intent(action)` method')
   }
 }
 
@@ -79,55 +74,72 @@ export class SaveState extends StateIntentBase {
   }
 
   static intent(action) {
-    return (event, _context, callback) => {
+    return async (event, context, callback) => {
       const { referenceId } = event.queryStringParameters
-      STATE_CLIENT.get(`api/v1/items/${referenceId}`)
-        .then(response => {
-          console.log('[BEARER]', 'received', response.data)
-          const state = response.data.Item
-          action(
-            event.context,
-            event.queryStringParameters,
-            event.body,
-            state,
-            result => {
-              STATE_CLIENT.put(`api/v1/items/${referenceId}`, {
-                ...result,
-                ReadAllowed: true
-              })
-                .then(data => {
-                  console.log('[BEARER]', 'success', data)
-                  callback(null, result)
-                })
-                .catch(e => {
-                  console.error('[BEARER]', 'error', e)
-                  callback(`Error : ${e}`)
-                })
-            }
-          )
+      const baseURL = event.context.bearerBaseURL || STATE_CLIENT.defaults.baseURL
+      try {
+        const response = await STATE_CLIENT.request({
+          method: 'get',
+          url: `api/v1/items/${referenceId}`,
+          baseURL
         })
-        .catch(response => {
-          action(
-            event.context,
-            event.queryStringParameters,
-            event.body,
-            {},
-            result => {
-              STATE_CLIENT.post(`api/v1/items`, {
-                ...result,
-                ReadAllowed: true
-              })
-                .then(data => {
-                  console.log('[BEARER]', 'success', data)
-                  callback(null, result)
-                })
-                .catch(e => {
-                  console.error('[BEARER]', 'error', e)
-                  callback(`Error : ${e}`)
-                })
+        console.log('[BEARER]', 'received', response.data)
+        const state = response.data.Item
+        action(event.context, event.queryStringParameters, event.body, state, result => {
+          STATE_CLIENT.request({
+            method: 'put',
+            url: `api/v1/items/${referenceId}`,
+            baseURL,
+            data: {
+              ...result,
+              ReadAllowed: true
             }
-          )
+          })
+            .then(data => {
+              console.log('[BEARER]', 'success', data)
+              callback(null, {
+                meta: {
+                  referenceId: referenceId
+                },
+                data: {
+                  ...result
+                }
+              })
+            })
+            .catch(e => {
+              console.error('[BEARER]', 'error', e)
+              callback(`Error : ${e}`)
+            })
         })
+      } catch (e) {
+        console.log(e)
+        action(event.context, event.queryStringParameters, event.body, {}, result => {
+          STATE_CLIENT.request({
+            method: 'post',
+            url: `api/v1/items`,
+            baseURL,
+            data: {
+              ...result,
+              ReadAllowed: true
+            }
+          })
+            .then((response: TStateData) => {
+              console.log('[BEARER]', 'success', response.data)
+              callback(null, {
+                meta: {
+                  referenceId: response.data.Item.referenceId
+                },
+                data: {
+                  ...result
+                }
+              })
+            })
+            .catch(e => {
+              console.error('[BEARER]', 'error', e)
+              callback(`Error : ${e}`)
+            })
+        })
+      }
     }
   }
 }
@@ -138,20 +150,27 @@ export class RetrieveState extends StateIntentBase {
   }
 
   static intent(action) {
-    return (event, _context, callback) => {
+    return (event, context, callback) => {
       const { referenceId } = event.queryStringParameters
+      const baseURL = event.context.bearerBaseURL || STATE_CLIENT.defaults.baseURL
 
-      STATE_CLIENT.get(`/api/v1/items/${referenceId}`)
+      STATE_CLIENT.request({
+        method: 'get',
+        url: `/api/v1/items/${referenceId}`,
+        baseURL
+      })
         .then(response => {
           if (response.data.error) {
             callback('No data found')
           } else {
             console.log('[BEARER]', 'data', response.data)
-            action(
-              event.context,
-              event.queryStringParameters,
-              response.data.Item,
-              state => callback(null, state)
+            action(event.context, event.queryStringParameters, response.data.Item, state =>
+              callback(null, {
+                meta: {
+                  referenceId: response.data.Item.referenceId
+                },
+                data: state
+              })
             )
           }
         })