+++ packages/transpiler/__test__/__fixtures__/pre/FullComponent.tsx
@@ -1,16 +1,11 @@
-import {
-  Component,
-  Intent,
-  BearerFetch,
-  IntentType,
-  BearerState
-} from '@bearer/core'
+import { Component, Intent, BearerFetch, IntentType, BearerState } from '@bearer/core'
 
 @Component({
   tag: 'full-component'
 })
 export class FullComponent {
   @Intent('ListRepositories') fetcher: BearerFetch
+  @RetrieveStateIntent() retrieve: BearerFetch
   @BearerState() attachedPullRequests: Array<any>
   @BearerState({ statePropName: 'goats' })
   ducks: Array<any>