+++ packages/transpiler/src/transformers/decorator-helpers.ts
@@ -17,24 +17,17 @@ export function hasName(node: ts.Decorator, tsDecoratorName: string) {
   return name(node) === tsDecoratorName
 }
 
-export function classDecoratedWithName(
-  node: ts.ClassDeclaration,
-  decoratorName: string
-): boolean {
+export function classDecoratedWithName(node: ts.ClassDeclaration, decoratorName: string): boolean {
   let hasComponentDecorator = false
   ts.forEachChild(node, n => {
     if (n.kind === ts.SyntaxKind.Decorator) {
-      hasComponentDecorator =
-        hasComponentDecorator || name(n as ts.Decorator) === decoratorName
+      hasComponentDecorator = hasComponentDecorator || name(n as ts.Decorator) === decoratorName
     }
   })
   return hasComponentDecorator
 }
 
-export function hasPropDecoratedWithName(
-  classNode: ts.ClassDeclaration,
-  decoratorName: string
-): boolean {
+export function hasPropDecoratedWithName(classNode: ts.ClassDeclaration, decoratorName: string): boolean {
   return Boolean(propDecoratedWithName(classNode, decoratorName).length)
 }
 
@@ -44,28 +37,19 @@ export function propDecoratedWithName(
 ): Array<ts.PropertyDeclaration | null> {
   const props: Array<ts.PropertyDeclaration | null> = []
   ts.forEachChild(node, node => {
-    if (
-      ts.isPropertyDeclaration(node) &&
-      hasDecoratorNamed(node as ts.PropertyDeclaration, decoratorName)
-    ) {
+    if (ts.isPropertyDeclaration(node) && hasDecoratorNamed(node as ts.PropertyDeclaration, decoratorName)) {
       props.push(node)
     }
   })
   return props
 }
 
-export function hasDecoratorNamed(
-  propDeclaration: ts.PropertyDeclaration,
-  name: string
-): boolean {
+export function hasDecoratorNamed(propDeclaration: ts.PropertyDeclaration, name: string): boolean {
   let has = false
   ts.forEachChild(propDeclaration, anode => {
     if (ts.isDecorator(anode)) {
       ts.forEachChild(anode, node => {
-        if (
-          ts.isCallExpression(node) &&
-          node.expression['escapedText'] === name
-        ) {
+        if (ts.isCallExpression(node) && node.expression['escapedText'] === name) {
           has = true
         }
       })
@@ -77,8 +61,7 @@ export function hasDecoratorNamed(
 export function using(node: ts.Node, decoratorName: string): boolean {
   let usedInCode = false
   function visit(node: ts.Node) {
-    if (ts.isDecorator(node))
-      usedInCode = usedInCode || hasName(node as ts.Decorator, decoratorName)
+    if (ts.isDecorator(node)) usedInCode = usedInCode || hasName(node as ts.Decorator, decoratorName)
     ts.forEachChild(node, visit)
   }