+++ packages/ui/src/components/Button/Button.stories.tsx
@@ -1,16 +1,7 @@
 import { storiesOf } from '@storybook/html'
 
 storiesOf('Button', module).add('kind', () => {
-  const kinds = [
-    'primary',
-    'secondary',
-    'success',
-    'danger',
-    'warning',
-    'info',
-    'light',
-    'dark'
-  ]
+  const kinds = ['primary', 'secondary', 'success', 'danger', 'warning', 'info', 'light', 'dark']
 
   return (
     '<table><tbody>' +