+++ packages/transpiler/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.22.4"></a>
+## [0.22.4](https://github.com/Bearer/bearer/compare/v0.22.2...v0.22.4) (2018-07-23)
+
+
+
+
+**Note:** Version bump only for package @bearer/transpiler
+
+<a name="0.22.3"></a>
+## [0.22.3](https://github.com/Bearer/bearer/compare/v0.22.3-0...v0.22.3) (2018-07-23)
+
+
+
+
+**Note:** Version bump only for package @bearer/transpiler
+
 <a name="0.20.1"></a>
 ## [0.20.1](https://github.com/Bearer/bearer/compare/v0.20.0...v0.20.1) (2018-07-20)